fix(codeql): address cpp-note-floats-includes-misc batch-2 alerts
- cpp/lossy-function-result-cast (speed.c): correct extract_channel
  return type from float→int; the function wraps speed_extract_score
  which returns int. Callers already used the result as an error code.
- cpp/lossy-function-result-cast (vif_tools.c): add explicit (int) casts
  on mirror() call sites; mirror() returns float but inputs are always
  integer-valued pixel indices so truncation is safe and intentional.
- cpp/constant-comparison (pdjson.c): remove redundant lower-bound
  `0x80 <= u` in utf8_seq_length(); the early `u < 0x80` guard above
  makes it always-true. Semantics unchanged.
- cpp/equality-on-floats: all 14 sites are intentional — sentinel checks
  (stored default vs option value, stored score idempotency, SVM integer
  class labels, normalised-zero fast-path) or bit-exact SIMD parity
  assertions. Added explanatory comments at each site; no epsilon
  introduced anywhere (would change scores or defeat correctness guards).
- cpp/loop-variable-changed (mkdirp.c): add comment documenting the
  intentional i++/i-- pattern used to skip duplicate path separators.
- cpp/include-non-header (10 sites): all deliberate white-box test
  single-TU includes; accepted, not changed.
- cpp/complex-block (7) + cpp/long-switch (1): SIMD kernels and upstream
  Y4M chroma upsampler; refactoring would risk bit-exactness or deviate
  from upstream; accepted, not changed.
- cpp/loop-variable-changed (svm.cpp vmaf_bench.c): upstream libsvm
  sentinel-scan + standard argv parsing; accepted with comments.
- cpp/commented-out-code (6): all illustrative/doc comments showing
  alternative algorithm modes; accepted, not changed.

+++ core/src/feature/speed.c
@@ -1298,8 +1298,8 @@ static int init_chroma(VmafFeatureExtractor *fex, enum VmafPixelFormat pix_fmt, 
     return 0;
 }
 
-static float extract_channel(SpeedChromaState *s, VmafPicture *ref_pic, VmafPicture *dist_pic,
-                             int channel, float *score)
+static int extract_channel(SpeedChromaState *s, VmafPicture *ref_pic, VmafPicture *dist_pic,
+                           int channel, float *score)
 {
     picture_copy(s->frame_buffer_ref, s->speed_state.float_stride, ref_pic, -128, ref_pic->bpc,
                  channel);
